Add uint8 version of utsnameStr
