+++ kernel_version.go
@@ -115,14 +115,3 @@ func CurrentKernelVersion() (uint32, error) {
 	}
 	return currentVersionUname()
 }
-
-func utsnameStr(in []int8) string {
-	out := make([]byte, len(in))
-	for i := 0; i < len(in); i++ {
-		if in[i] == 0 {
-			break
-		}
-		out = append(out, byte(in[i]))
-	}
-	return string(out)
-}
